Use for-each-group to retrieve authors
The previous approach (first retrieving authors in a variable, and then filter
publications with the author names) did not work because XPath dynamic
evaluation is not possible with XSLT 1.0 & 2.0.

Instead, we used for-each-group to retrieve authors. By using this
approach, the context is preserved so we can retrieve the root element
(in the previous approach we had to store the root in a variable), and
filter the publications with the grouping key (which is not considered
dynamic evaluation).