publication - position
Publications are not grouped by year using for-each-group anymore.
Instead, we display a new table header when the current publication's
year differs from the previous one.

The for-each-group made our task of retrieving the publication's
position difficult, because the position was relative to the year group and
not the total of publications for the current author. So we had:

2008
2 title
1 title
2007
1 title

Instead of:

2008
3 title
2 title
2007
1 title

By using a simple for-each instead of the for-each-group, the use of the
position() function is now possible and the output is the intended one.