Fix year header
To compare the current publication's year and the one from the previous
publication, we used the preceding-sibling axis. However, it does not
work as intended on sorted elements. This axis returns the preceding
sibling as if the elements were not sorted.

In order to bypass this limitation, we sorted the publication's years in
a variable, and used it to retrieve the previous publication's year with
the help of the position() function.

N.B. We tried to use $variable[position()] directly, but it did not seem
to work. Instead, we stored the position() in an intermediate variable.